+++ SpeechResponder/Properties/SpeechResponder.Designer.cs
@@ -19,7 +19,7 @@ namespace EddiSpeechResponder.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "18.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     public class SpeechResponder {
@@ -309,6 +309,15 @@ namespace EddiSpeechResponder.Properties {
         public static string localizedAnd {
             get {
                 return ResourceManager.GetString("localizedAnd", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to or.
+        /// </summary>
+        public static string localizedOr {
+            get {
+                return ResourceManager.GetString("localizedOr", resourceCulture);
             }
         }